Updated Inventories to use Primary Key. (#65)
* Updated Inventories to use Primary Key.

This change simplifies inventories by setting InventoryId as the primary key.  Unstead of saying  PlayerInventory, we set the InventoryId for those entries to the PlayerId. We're also removing the PlayerId column from the table as its no longer needed. To obtain if an entry is a player inventory, we still have the IsPlayerInventory bool.

This change is a stepping stone towards improved bulk updating of inventories as the current implementation updates all inventories one at a time rather than all at once.

* Hotfix.